Update session id in redirect call.
