+++ cartridges/int_bolt_core/cartridge/scripts/utils/oauthUtils.js
@@ -12,7 +12,8 @@ var HttpResult = require('dw/svc/Result');
 var JWTUtils = require('int_bolt_core/cartridge/scripts/utils/jwtUtils');
 var BoltHttpUtils = require('int_bolt_core/cartridge/scripts/services/utils/httpUtils');
 var BoltPreferences = require('int_bolt_core/cartridge/scripts/services/utils/preferences');
-var CommonUtils = require('int_bolt_core/cartridge/scripts/utils/commonUtils');
+var LogUtils = require('int_bolt_core/cartridge/scripts/utils/boltLogUtils');
+var log = LogUtils.getLogger('Login');
 
 const OpenIdEndpoint = '/.well-known/openid-configuration';
 const BoltProviderID = 'Bolt';
@@ -25,7 +26,7 @@ const BoltProviderID = 'Bolt';
  * @param {string} orderUUID - the created order UUID if provided
  * @returns {Object} result
  */
-exports.oauthLoginOrCreatePlatformAccount = function (code, scope, orderId, orderUUID, boltOrderId) {
+exports.oauthLoginOrCreatePlatformAccount = function (code, scope, orderId, orderUUID) {
   // step 1: fetch openID configuration from Bolt
   var { clientID, clientSecret, providerID, boltAPIbaseURL } = getOAuthConfiguration();
   var openIDConfigResponse = BoltHttpUtils.restAPIClient('GET', '', '', 'none', boltAPIbaseURL+OpenIdEndpoint);
@@ -63,10 +64,6 @@ exports.oauthLoginOrCreatePlatformAccount = function (code, scope, orderId, orde
       CustomerMgr.loginExternallyAuthenticatedCustomer(providerID, platformAccountID, false);
     });
 
-    // update session id in Bolt since dwsid changed after shopper login
-    if (boltOrderId.value) {
-      putSFCCObject(orderId, boltOrderId, clientID);
-    }
   } else {
     return oauthErrorResponse('Platform account customer profile credentials disabled: ' + platformAccountID);
   }
@@ -79,29 +76,6 @@ exports.oauthLoginOrCreatePlatformAccount = function (code, scope, orderId, orde
     email: externalProfile.email,
   };
 };
-
-/**
- * Communicates with Bolt db Endpoint to put the cart object
- * @param {string} orderID - SFCC order id
- * @param {string} boltOrderId -  Bolt order id
- * @param {string} clientID - client id for the oauth workflow, should be the same as merchant publishable key
- */
-function putSFCCObject (orderId, boltOrderId, clientId) {
-  const endpoint = '/sfcc_objects';
-  var dwsid = CommonUtils.getDwsidCookie();
-
-  const request = {
-    publishable_key: clientId,
-    sfcc_order_id: orderId,
-    session_id: dwsid,
-    order_id: boltOrderId,
-  };
-
-  var serviceResponse = BoltHttpUtils.restAPIClient('POST', endpoint, JSON.stringify(request), '', '');
-  if (!serviceResponse || serviceResponse.status == HttpResult.ERROR) {
-      log.error('Failed to update dwid to Bolt.')
-  }
-}
 
 /**
  * Return OAuth configuration
